now vulkan

+++ loo/runtime/looreflect/include/looreflect/declenum.h
@@ -13,7 +13,7 @@ namespace looreflect {
 			unsigned flags,
 			const char* name,
 			const char* tagName,
-			int value,
+			uint64_t value,
 			int index)noexcept
 			:m_type (type)
 			, m_flags (flags)
@@ -23,10 +23,11 @@ namespace looreflect {
 			, m_index (index) {
 		}
 
-		int value ()const noexcept {
+		uint64_t value ()const noexcept {
 			return m_value;
 		}
-		int index ()const noexcept {
+
+		uint64_t index ()const noexcept {
 			return m_index;
 		}
 		std::size_t flags () const noexcept {
@@ -46,7 +47,8 @@ namespace looreflect {
 		}
 	public:
 		int m_index;
-		int m_value;
+		uint64_t m_value;
+		
 		const char* m_name;//not full name
 		const char* m_tagName;
 		std::size_t m_flags;
@@ -142,7 +144,7 @@ namespace looreflect {
 			return nullptr;
 		}
 
-		LooEnumConstant const* enum_constant_by_value (int value)const noexcept {
+		LooEnumConstant const* enum_constant_by_value (uint64_t value)const noexcept {
 			for (auto& enumvalue : values ()) {
 				if (value == enumvalue.value ()) {
 					return &enumvalue;
